pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pdf-ingest"
version = "0.1.0"
description = "PDF document ingestion CLI for evidence-traceable Obsidian literature notes"
requires-python = ">=3.11"
license = {text = "MIT"}

dependencies = [
    "pymupdf>=1.26.0,<1.28.0",
    "typer>=0.9.0",
    "pydantic>=2.0.0",
    "pillow>=10.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "jsonschema>=4.0.0",
]

[project.scripts]
ingest_pdf = "ingest.cli:main"
ingest_qa_regression = "ingest.qa_regression:main"

[tool.hatch.build.targets.wheel]
packages = ["src/ingest"]

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/schemas",
    "/prompts",
]
